[699] Only expose RepresentationMetadata in most of the GraphQL Schema
In almost all locations where we used to return a full
Representation (or a concrete Diagram), only expose the new
RepresentationMetadata type. The only exception is in the
subscription, where we actually *need* the full representation's
content.

This ensures the underlying's representation content (e.g. the
nodes/edges/etc. on a diagram) can not be accessed by the frontend
except in the specific case where it is needed.

This change is purely on the GraphQL Schema level and does not change
anything to which Java objects are loaded/manipulated on the backend.
It just makes sure that except for the case of refresh events, only
the metadata fields can be requested by and returned to the frontend.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/699
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>